uninstall, direct access cleanup

+++ uninstall.php
@@ -1,37 +1,27 @@
 <?php
-
 /**
+ * Uninstall functionality for JobCapturePro plugin.
+ *
  * Fired when the plugin is uninstalled.
  *
- * When populating this file, consider the following flow
- * of control:
- *
- * - This method should be static
- * - Check if the $_REQUEST content actually is the plugin name
- * - Run an admin referrer check to make sure it goes through authentication
- * - Verify the output of $_GET makes sense
- * - Repeat with other user roles. Best directly by using the links/query string parameters.
- * - Repeat things for multisite. Once for a single site in the network, once sitewide.
- *
- * This file may be updated more in future version of the Boilerplate; however, this is the
- * general skeleton and outline for how the file should work.
- *
- * For more information, see the following discussion:
- * https://github.com/tommcfarlin/WordPress-Plugin-Boilerplate/pull/123#issuecomment-28541913
- *
- * @package    JobCapturePro
- * @subpackage JobCapturePro/uninstall
+ * @package JobCapturePro
+ * @since   1.0.0
  */
 
 // If uninstall not called from WordPress, then exit.
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
-}
-
 /**
  * Clean up plugin data on uninstall
+ *
+ * @since 1.0.0
  */
 function jobcapturepro_uninstall() {
+	// Check user permissions
+	if ( ! current_user_can( 'delete_plugins' ) ) {
+		return;
+	}
+
 	// Remove plugin options
 	delete_option( 'jobcapturepro_options' );
